+++ lib/core/apis/b2b-Topup.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   generateOAuthToken,
   encryptSecurityCredential,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 
 /**
  * B2B Account Top-Up - Use this API to load funds to a B2B shortcode for disbursement.
