+++ lib/core/apis/b2c-Request.js
@@ -1,10 +1,10 @@
 import axios from "axios";
-import { CommandIDs } from "./utils/constants.js";
+import { CommandIDs } from "../utils/constants.js";
 import {
   encryptSecurityCredential,
   generateOAuthToken,
   originatorID,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 
 /**
  * B2C Payment Request
