+++ lib/core/apis/balance-Query.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   encryptSecurityCredential,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 /**
  * AccountBalance - Use this API to enquire the balance on an M-Pesa BuyGoods (Till Number).
  * Simultaneously generates an OAuth token using the consumer key and secret.
