+++ lib/core/apis/business-Paybill.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   encryptSecurityCredential,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 
 /**
  * businessPaybillApi - Use this API to pay bills directly from your business account to a pay bill number or store.
