+++ lib/core/apis/c2b-Register.js
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { generateOAuthToken } from "./utils/helpers.js";
-import { responseTypes } from "./utils/constants.js";
+import { generateOAuthToken } from "../utils/helpers.js";
+import { responseTypes } from "../utils/constants.js";
 
 /**
  * C2B Register URL - NOT SIMULATION
