+++ lib/core/apis/c2b-Simulate.js
@@ -1,4 +1,4 @@
-import { generateOAuthToken } from "./utils/helpers.js";
+import { generateOAuthToken } from "../utils/helpers.js";
 import axios from "axios";
 
 /**
