+++ lib/core/apis/mpesa-Query.js
@@ -1,7 +1,7 @@
 import {
   generateMpesaCredentials,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 import axios from "axios";
 /**
  * Lipa Na M-Pesa Query Request - Use this API to check the status of a Lipa Na M-Pesa Online Payment.
