+++ lib/core/apis/mpesa-Simulate.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   generateMpesaCredentials,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 /**
  * Lipa Na M-Pesa Online Payment - Use this API to initiate an online payment on behalf of a customer.
  * @name mpesaSimulateApi
