+++ lib/core/apis/qr-Generate.js
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { generateOAuthToken } from "./utils/helpers.js";
-import { trxCodeTypes } from "./utils/constants.js";
+import { generateOAuthToken } from "../utils/helpers.js";
+import { trxCodeTypes } from "../utils/constants.js";
 
 /**
  * generateQrCodeApi - Use this API to create a dynamic QR code for a specified transaction.
