+++ lib/core/apis/reversals.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   encryptSecurityCredential,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 /**
  * Reversal Request - Use this API to reverse an M-Pesa transaction C2B.
  * @name reversalsApi
