+++ lib/core/apis/tax-Remittance.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import {
   generateOAuthToken,
   encryptSecurityCredential,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 
 /**
  * Tax Remittance API - Use this API to remit tax to the Kenya Revenue Authority (KRA).
