+++ lib/core/apis/transaction-Status.js
@@ -1,7 +1,7 @@
 import {
   encryptSecurityCredential,
   generateOAuthToken,
-} from "./utils/helpers.js";
+} from "../utils/helpers.js";
 import axios from "axios";
 /**
  * Transaction Status Request - Use this API to check the status of a transaction on M-Pesa.
