+++ lib/core/utils/constants.js
