+++ lib/core/utils/helpers.js
